+++ app/src/main/java/shub39/momentum/viewmodels/ProjectViewModel.kt
@@ -25,14 +25,18 @@ import androidx.media3.common.MediaItem
 import androidx.media3.exoplayer.ExoPlayer
 import androidx.media3.exoplayer.ExoPlayer.Builder
 import androidx.media3.exoplayer.ExoPlayer.REPEAT_MODE_ALL
+import io.github.vinceglb.filekit.PlatformFile
+import io.github.vinceglb.filekit.dialogs.toAndroidUri
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.flow.stateIn
@@ -40,10 +44,12 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.koin.core.annotation.KoinViewModel
+import shub39.momentum.core.data_classes.MontageConfig
+import shub39.momentum.core.data_classes.toMontageConfig
+import shub39.momentum.core.data_classes.toMontageOptions
 import shub39.momentum.core.enums.VideoAction
 import shub39.momentum.core.interfaces.AlarmScheduler
 import shub39.momentum.core.interfaces.FaceDetector
-import shub39.momentum.core.interfaces.MontageConfigPrefs
 import shub39.momentum.core.interfaces.MontageMaker
 import shub39.momentum.core.interfaces.MontageState
 import shub39.momentum.core.interfaces.ProjectRepository
@@ -58,7 +64,6 @@ class ProjectViewModel(
     private val montageMaker: MontageMaker,
     private val faceDetector: FaceDetector,
     private val repository: ProjectRepository,
-    private val montageConfigPrefs: MontageConfigPrefs,
     private val scheduler: AlarmScheduler,
     private val imageHandler: ImageHandler,
 ) : ViewModel() {
@@ -109,7 +114,8 @@ class ProjectViewModel(
             is ProjectAction.OnUpsertDay ->
                 viewModelScope.launch {
                     if (action.isNewImage) {
-                        val faceData = faceDetector.getFaceDataFromUri(action.day.image.toUri())
+                        val uri = PlatformFile(action.day.image).toAndroidUri()
+                        val faceData = faceDetector.getFaceDataFromUri(uri)
                         val copiedImageUri = imageHandler.copyImageToAppData(action.day)
 
                         Log.d("ProjectViewModel", "faceData : $faceData")
@@ -181,18 +187,9 @@ class ProjectViewModel(
 
             is ProjectAction.OnEditMontageConfig -> {
                 viewModelScope.launch {
-                    montageConfigPrefs.setFpi(action.config.framesPerImage)
-                    montageConfigPrefs.setFps(action.config.framesPerSecond)
-                    montageConfigPrefs.setShowDate(action.config.showDate)
-                    montageConfigPrefs.setShowMessage(action.config.showMessage)
-                    montageConfigPrefs.setFont(action.config.font)
-                    montageConfigPrefs.setDateStyle(action.config.dateStyle)
-
-                    montageConfigPrefs.setVideoQuality(action.config.videoQuality)
-                    montageConfigPrefs.setStabilizeFaces(action.config.stabilizeFaces)
-                    montageConfigPrefs.setBackgroundColor(action.config.backgroundColor)
-                    montageConfigPrefs.setWaterMark(action.config.waterMark)
-                    montageConfigPrefs.setCensorPref(action.config.censorFaces)
+                    _state.value.project?.id?.let { projectId ->
+                        updateMontageConfig(projectId = projectId, config = action.config)
+                    }
                 }
             }
 
@@ -210,8 +207,12 @@ class ProjectViewModel(
                     }
                 }
 
-            ProjectAction.OnResetMontagePrefs ->
-                viewModelScope.launch { montageConfigPrefs.resetPrefs() }
+            is ProjectAction.OnResetMontagePrefs ->
+                viewModelScope.launch {
+                    _state.value.project?.id?.let { projectId ->
+                        updateMontageConfig(projectId = projectId, config = MontageConfig())
+                    }
+                }
 
             ProjectAction.OnStartFaceScan ->
                 viewModelScope.launch {
@@ -230,7 +231,8 @@ class ProjectViewModel(
                                     )
                                 }
 
-                                val faceData = faceDetector.getFaceDataFromUri(day.image.toUri())
+                                val uri = PlatformFile(day.image).toAndroidUri()
+                                val faceData = faceDetector.getFaceDataFromUri(uri)
                                 repository.upsertDay(day.copy(faceData = faceData))
                             }
                         }
@@ -244,6 +246,14 @@ class ProjectViewModel(
             ProjectAction.OnResetScanState -> {
                 _state.update { it.copy(scanState = ScanState.Idle) }
             }
+        }
+    }
+
+    private suspend fun updateMontageConfig(projectId: Long, config: MontageConfig) {
+        repository.upsertMontageOptions(config.toMontageOptions(projectId))
+
+        repository.getMontageOptionsByProjectId(projectId)?.toMontageConfig()?.let { config ->
+            _state.update { it.copy(montageConfig = config) }
         }
     }
 
@@ -269,7 +279,8 @@ class ProjectViewModel(
     private suspend fun processDays() {
         state.value.days.forEach { day ->
             if (day.faceData == null) {
-                val faceData = faceDetector.getFaceDataFromUri(day.image.toUri())
+                val uri = PlatformFile(day.image).toAndroidUri()
+                val faceData = faceDetector.getFaceDataFromUri(uri)
                 repository.upsertDay(day.copy(faceData = faceData))
             }
         }
@@ -277,99 +288,17 @@ class ProjectViewModel(
 
     private fun observeConfig() =
         viewModelScope.launch {
-            montageConfigPrefs
-                .getFpiFlow()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(framesPerImage = pref))
-                    }
-                }
-                .launchIn(this)
+            // get montage config when project changes
+            _state
+                .map { it.project?.id }
+                .distinctUntilChanged()
+                .onEach { projectId ->
+                    projectId?.let {
+                        val config =
+                            repository.getMontageOptionsByProjectId(it)?.toMontageConfig()
+                                ?: MontageConfig()
 
-            montageConfigPrefs
-                .getFpsFlow()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(framesPerSecond = pref))
-                    }
-                }
-                .launchIn(this)
-
-            montageConfigPrefs
-                .getVideoQualityFlow()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(videoQuality = pref))
-                    }
-                }
-                .launchIn(this)
-
-            montageConfigPrefs
-                .getBackgroundColorFlow()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(backgroundColor = pref))
-                    }
-                }
-                .launchIn(this)
-
-            montageConfigPrefs
-                .getWaterMarkFlow()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(waterMark = pref))
-                    }
-                }
-                .launchIn(this)
-
-            montageConfigPrefs
-                .getShowDateFlow()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(showDate = pref))
-                    }
-                }
-                .launchIn(this)
-
-            montageConfigPrefs
-                .getShowMessageFlow()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(showMessage = pref))
-                    }
-                }
-                .launchIn(this)
-
-            montageConfigPrefs
-                .getFontFlow()
-                .onEach { pref ->
-                    _state.update { it.copy(montageConfig = it.montageConfig.copy(font = pref)) }
-                }
-                .launchIn(this)
-
-            montageConfigPrefs
-                .getDateStyleFlow()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(dateStyle = pref))
-                    }
-                }
-                .launchIn(this)
-
-            montageConfigPrefs
-                .getStabilizeFacesFlow()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(stabilizeFaces = pref))
-                    }
-                }
-                .launchIn(this)
-
-            montageConfigPrefs
-                .getCensorPref()
-                .onEach { pref ->
-                    _state.update {
-                        it.copy(montageConfig = it.montageConfig.copy(censorFaces = pref))
+                        _state.update { state -> state.copy(montageConfig = config) }
                     }
                 }
                 .launchIn(this)