+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.7.0
+- Added Camera Option for adding Pictures in Days
+- Montage Preferences are saved per project
+- Miscellaneous UI Improvements
+
 ## 1.6.1
 - Miscellaneous Updates and Improvements
 - Notifications open App on click